% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_csv_function.R
\name{open_csv}
\alias{open_csv}
\title{A function for opening csvs that are saved by SSMS 2012.}
\usage{
open_csv(filename, app = "SSMS", header = T)
}
\arguments{
\item{filename}{Path on your file system to the saved-down file.}

\item{app}{Keyword for the application that generated your CSV.  Defaults to SSMS.}

\item{header}{Boolean for whether your file includes headers.}
}
\description{
This function lets you open up .csv files that are saved from SQL Server Management Studio 2012.
}
\examples{
open_csv()
}
\keyword{CSV}
\keyword{Microsoft}
\keyword{SQL}
\keyword{SQL,}
\keyword{SSMS,}
\keyword{Server,}

